+++ list_rental_by_id.rb
@@ -1,5 +1,4 @@
 def list_rentals_for_person
-
   puts 'Enter person id:'
 
   id = gets.chomp.to_i
@@ -11,9 +10,7 @@ def list_rentals_for_person
   if person and person.rental.length.positive?
 
     person.rental.each do |r|
-
       puts "Person Name: #{person.name}, Book: #{r.book.title} by #{r.book.author}, Date: #{r.date}"
-
     end
 
   else
@@ -21,6 +18,4 @@ def list_rentals_for_person
     puts 'No rentals found'
 
   end
-
 end
-
