+++ person.rb
@@ -73,15 +73,3 @@ class TrimmerDecorator < BaseDecorator
     end
   end
 end
-
-# person = Person.new(2, 'maximilianus', 45)
-
-# puts person.correct_name
-
-# capitalized_person = CapitalizeDecorator.new(person)
-
-# puts capitalized_person.correct_name
-
-# capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
-
-# puts capitalized_trimmed_person.correct_name
