+++ student.rb
@@ -1,7 +1,6 @@
 require_relative 'person'
 
 class Student < Person
-
   attr_accessor :classroom
 
   def initialize(id, age, name, classroom, parent_permission)
@@ -12,20 +11,11 @@ class Student < Person
 
   def play_hooky
     '¯\(ツ)/¯'
-
   end
-
-
 
   def add_classroom=(classroom)
-
     @classroom = classroom
 
-
-
     classroom.students.push(self) unless classroom.students.include?(self)
-
   end
-
 end
-
